feat: db에서 가져온 데이터를 dto로 매핑하는 기능 구현
