+++ README.md
@@ -12,7 +12,7 @@
 - [ ] 게임 플레이 이력 조회 API 구현
   - [x] 게임 플레이 이력에 대한 dto 구성
   - [x] 게임 플레이 이력을 가져오는 기능 구현
-  - [ ] db에서 가져온 데이터를 dto로 매핑하는 기능 구현
+  - [x] db에서 가져온 데이터를 dto로 매핑하는 기능 구현
   - [ ] get 요청을 받는 controller 구현
 - [ ] 기존 기능 수정 - 출력 방식 수정
   - [ ] 플레이의 중간 과정을 출력하는 로직 제거
